+++ src/betterdiscord/api/fetch.ts
@@ -1,125 +1,83 @@
+import {dryReadableStream, hydrateReadableStream, type DriedAbortSignal, type DriedResponse, type NativeRequestInit, type NativeRequestMethod} from "@common/native-fetch.ts";
 import Remote from "../polyfill/remote";
 
-const redirects = new Set(["manual", "follow"]);
+const MAX_DEFAULT_REDIRECTS = 20;
+const DEFAULT_TIMEOUT = 3000;
 const methods = new Set(["GET", "PUT", "POST", "DELETE", "PATCH", "OPTIONS", "HEAD", "CONNECT", "TRACE"]);
-const bodylessStatusCodes = new Set([101, 204, 205, 304]);
 
+function dryAbortSignal(signal: AbortSignal): DriedAbortSignal {
+    return {
+        aborted: () => signal.aborted,
+        reason: () => signal.reason,
+        addListener: (onAbort) => {
+            if (signal.aborted) {
+                try {
+                    onAbort();
+                }
+                finally {
+                    // eslint-disable-next-line no-unsafe-finally
+                    return () => {};
+                }
+            }
 
-class FetchResponse extends Response {
+            const listener = () => onAbort();
+            signal.addEventListener("abort", listener, {once: true});
+            return () => void signal.removeEventListener("abort", listener);
+        }
+    };
+}
 
-    private _options: Partial<FetchData & {status?: number;}>;
+class HydratingResponse extends Response {
+    #dried;
 
-    constructor(options: Partial<FetchData & {status?: number;}>) {
-        super(bodylessStatusCodes.has(options.status ?? 0) ? null : options.content as any, {
-            headers: new Headers(options.headers),
-            method: options.method ?? "GET",
-            body: options.content,
-            ...options
+    constructor(dried: DriedResponse) {
+        super(dried.body ? hydrateReadableStream(dried.body) : null, {
+            headers: dried.headers,
+            status: dried.status,
+            statusText: dried.statusText
         });
 
-        this._options = options;
+        this.#dried = dried;
     }
 
-    get url() {return this._options.url!;}
-    get redirected() {return this._options.redirected!;}
+    override get url() {
+        return this.#dried.url;
+    }
+    override get redirected() {
+        return this.#dried.redirected;
+    }
 }
 
-const convertSignal = (signal: AbortSignal) => {
-    const listeners = new Set<() => void>();
+async function fetch(input: string | URL | Request, init?: NativeRequestInit): Promise<Response> {
+    let request: Request;
+    if (input instanceof Request) {
+        request = input;
+    }
+    else if (input instanceof URL) {
+        request = new Request(input.href, init);
+    }
+    else {
+        request = new Request(input, init);
+    }
 
-    signal.addEventListener("abort", () => {
-        listeners.forEach(l => l());
+    const driedResponse = await Remote.nativeFetch({
+        url: request.url,
+
+        body: request.body ? dryReadableStream(request.body) : null,
+
+        headers: Object.fromEntries(request.headers),
+
+        keepalive: request.keepalive,
+        method: methods.has(request.method) ? request.method as NativeRequestMethod : "GET",
+        redirect: request.redirect,
+        signal: request.signal ? dryAbortSignal(request.signal) : null,
+
+        timeout: init?.timeout ?? DEFAULT_TIMEOUT,
+        maxRedirects: init?.maxRedirects ?? MAX_DEFAULT_REDIRECTS,
+        rejectUnauthorized: init?.rejectUnauthorized ?? true
     });
 
-    return {
-        addEventListener(_: any, listener: () => void) {
-            listeners.add(listener);
-        }
-    } as AbortSignal;
-};
-
-/**
- * @typedef {Object} FetchOptions
- * @property {"GET" | "PUT" | "POST" | "DELETE" | "PATCH" | "OPTIONS" | "HEAD" | "CONNECT" | "TRACE"} [method] - Request method.
- * @property {Record<string, string>} [headers] - Request headers.
- * @property {"manual" | "follow"} [redirect] - Whether to follow redirects.
- * @property {number} [maxRedirects] - Maximum amount of redirects to be followed.
- * @property {AbortSignal} [signal] - Signal to abruptly cancel the request
- * @property {Uint8Array | string} [body] - Defines a request body. Data must be serializable.
- * @property {number} [timeout] - Request timeout time.
- */
-
-// TODO: de-dup with preload and make @common/types
-interface FetchOptions {
-    method: "GET" | "PUT" | "POST" | "DELETE" | "PATCH" | "OPTIONS" | "HEAD" | "CONNECT" | "TRACE";
-    headers: Record<string, string>;
-    redirect: "manual" | "follow";
-    maxRedirects: number;
-    signal: AbortSignal;
-    body: Uint8Array | string;
-    timeout: number;
+    return new HydratingResponse(driedResponse);
 }
 
-interface FetchData extends FetchOptions {
-    content: Buffer[] | Buffer;
-    headers: Record<string, any>;
-    statusCode: number;
-    url: string;
-    statusText: string;
-    redirected: boolean;
-    signal: AbortSignal;
-}
-
-
-/**
- * @param {string} url
- * @param {FetchOptions} options
- * @returns {Promise<FetchResponse>}
- */
-export default function fetch(url: string, options: Partial<FetchOptions> = {}) {
-    return new Promise((resolve, reject) => {
-        const data: Partial<FetchData> = {};
-
-        if (typeof options.headers === "object") {
-            data.headers = options.headers instanceof Headers ? Object.fromEntries(options.headers.entries()) : options.headers;
-        }
-
-        if (typeof options.redirect === "string" && redirects.has(options.redirect)) data.redirect = options.redirect;
-        if (typeof options.body === "string" || options.body instanceof Uint8Array) data.body = options.body;
-        if (typeof options.method === "string" && methods.has(options.method)) data.method = options.method;
-        if (typeof options.maxRedirects === "number") data.maxRedirects = options.maxRedirects;
-        if (typeof options.timeout === "number") data.timeout = options.timeout;
-
-        if (options.signal instanceof AbortSignal) data.signal = convertSignal(options.signal);
-
-        let ctx;
-        try {
-            ctx = Remote.nativeFetch(url, data);
-        }
-        catch (error) {
-            return reject(error);
-        }
-
-        ctx.onError(error => {
-            reject(error);
-        });
-
-        ctx.onComplete(() => {
-            try {
-                const resultData = ctx.readData();
-
-                const req = new FetchResponse({
-                    method: options.method ?? "GET",
-                    status: resultData?.statusCode,
-                    ...options,
-                    ...resultData
-                });
-
-                resolve(req);
-            }
-            catch (error) {
-                reject(error);
-            }
-        });
-    });
-}
+export default fetch;