+++ src/betterdiscord/ui/settings/addoncard.tsx
@@ -13,10 +13,10 @@ import Switch from "./components/switch";
 import Modals from "@ui/modals";
 
 import {CircleDollarSignIcon, CircleHelpIcon, PlugIcon, GithubIcon, GlobeIcon, HeartHandshakeIcon, PaletteIcon, PencilIcon, SettingsIcon, ShieldAlertIcon, Trash2Icon} from "lucide-react";
-import {getStore} from "@webpack";
+import {getByKeys} from "@webpack";
 import type {Addon} from "@modules/addonmanager";
 import type {MouseEvent, ReactNode} from "react";
-import type AddonManager from "@modules/addonmanager"; // eslint-disable-line no-duplicate-imports
+import type AddonManager from "@modules/addonmanager";  
 
 const {useCallback, useMemo} = React;
 
@@ -48,11 +48,11 @@ const LayerManager = {
     }
 };
 
-const UserStore = getStore("UserStore");
-const ChannelStore = getStore("ChannelStore");
-const PrivateChannelActions = DiscordModules.PrivateChannelActions;
-const ChannelActions = DiscordModules.ChannelActions;
-const getString = (value: string | {toString(): string;}) => typeof value == "string" ? value : value.toString();
+const UserStore = getByKeys<{getCurrentUser(): {id: string;};}>(["getCurrentUser"]);
+const ChannelStore = getByKeys<{getDMFromUserId(id: string): string;}>(["getDMFromUserId"]);
+const PrivateChannelActions = getByKeys<{openPrivateChannel(me: string, them: string): void;}>(["openPrivateChannel"]);
+const ChannelActions = getByKeys<{selectPrivateChannel(id: string): void;}>(["selectPrivateChannel"]);
+const getString = (value: string | {toString(): string;}) => typeof value == "string" ? value : value?.toString?.() || "";
 
 function makeButton(title: string, children: ReactNode, action?: () => void, {isControl = false, danger = false, disabled = false} = {}) {
     const ButtonType = isControl ? "button" : "div";